First prototype of the data analysis module of RoadTripper
This is a set of functionalities that allow to clean the data of
potholes alerts and jams alerts (we use k-means clustering) and then it
stores averything in a new table (.cvs file) to store it in the server.
Additionally, the public transport routes are analyzed to determine
which ones are affected by potholes, and also, Voronoi polygons are
created from the jams alerts to determine congestion areas in the city
(also using k-means). The processed data is stored in the CSV files and
in the .mat files.